Refactoring: old module `cfx` becomes `model` while introducing a new `cfx`.
