+++ crypto-forex/src/cfx.rs
@@ -1,81 +1,60 @@
-//! Module that retrieves information from the cryptocurrencies database. Fundamentally, it contains
-//! the `CryptoPlatform` struct that represents a cryptocurrency platform used by other, the
-//! `CryptoMap` struct that describes all the characteristics of a specific cryptocurrency and the
-//! `CryptoListing` struct provides accurate and timely data for the cryptoasset.
+use sqlx::PgPool;
+use std::net::TcpListener;
+use thiserror::Error;
 
-use chrono::prelude::*;
-use rust_decimal::prelude::*;
-use sqlx::FromRow;
+use crate::route::{crypto_listing, crypto_map, health_check};
+use crate::{config, database};
+use actix_web::{
+    dev::Server,
+    web::{self, Data},
+    App, HttpServer,
+};
 
-#[derive(Debug, FromRow)]
-pub struct CryptoPlatform {
-    pub id: i32,
-    pub platform: i32,
-    pub token_address: String,
+pub struct CfxServer {
+    port: u16,
+    server: Server,
 }
 
-#[derive(Debug, FromRow)]
-pub struct CryptoMap {
-    pub id: i32,
-    pub name: String,
-    pub symbol: String,
-    pub slug: String,
-    pub rank: i32,
-    pub is_active: bool,
-    pub first_historical_data: DateTime<Utc>,
-    pub last_historical_data: DateTime<Utc>,
-    pub platform: Option<i32>,
+impl CfxServer {
+    pub async fn build(config: config::Configuration) -> Result<Self, std::io::Error> {
+        let db_pool = database::get_connection_pool(&config.database);
+
+        let address = format!("{}:{}", config.application.host, config.application.port);
+        let listener = TcpListener::bind(&address)?;
+        let port = listener.local_addr().unwrap().port();
+        let server = setup(listener, db_pool)?;
+
+        Ok(Self { port, server })
+    }
+
+    pub fn port(&self) -> u16 {
+        self.port
+    }
+
+    pub async fn run(self) -> Result<(), std::io::Error> {
+        self.server.await
+    }
 }
 
-#[derive(Debug, FromRow)]
-pub struct CryptoListing {
-    /// The CoinMarketCap's `id`.
-    pub id: i32,
-    /// Number of market pairs across all exchanges trading each currency.
-    pub num_market_pairs: i32,
-    pub tags: Vec<String>,
-    /// Approximation of the maximum amount of coins that will ever exist in the lifetime
-    /// of the currency.
-    pub max_supply: Option<Decimal>,
-    /// The amount of coins that are circulating in the market and are in public hands. It is
-    /// analogous to the flowing shares in the stock market.
-    pub circulating_supply: Decimal,
-    /// Approximate total amount of coins in existence right now (minus any coins that have been
-    /// verifiably burned).
-    pub total_supply: Decimal,
-    pub platform: Option<i32>,
-    /// CoinMarketCap's market cap rank as outlined in [their methodology](https://coinmarketcap.com/methodology/).
-    /// Cryptocurrencies are listed by `cmc_rank` by default.
-    pub cmc_rank: i32,
-    pub quote: String,
+#[derive(Error, Debug)]
+pub enum CfxError {
+    #[error("Issues loading configuration")]
+    CfgError(#[from] ::config::ConfigError),
+    #[error("Issues with I/O operations")]
+    IoError(#[from] std::io::Error),
+}
 
-    /// Latest average trade price across markets.
-    pub price: Decimal,
-    /// A measure of how much of a cryptocurrency was traded in the last 24 hours.
-    pub volume_24h: Decimal,
-    pub volume_change_24h: Decimal,
-    /// 1 hour trading price percentage change for each currency.
-    pub percent_change_1h: Decimal,
-    /// 24 hour trading price percentage change for each currency.
-    pub percent_change_24h: Decimal,
-    /// 7 day trading price percentage change for each currency.
-    pub percent_change_7d: Decimal,
-    pub percent_change_30d: Decimal,
-    pub percent_change_60d: Decimal,
-    pub percent_change_90d: Decimal,
-    /// The total market value of a cryptocurrency's circulating supply. It is analogous to the
-    /// free-float capitalization in the stock market.
-    ///
-    /// `Market Cap = Current Price x Circulating Supply`
-    ///
-    /// (see [details](https://coinmarketcap.com/methodology/))
-    pub market_cap: Decimal,
-    pub market_cap_dominance: Decimal,
-    /// The market cap if the max supply was in circulation.
-    ///
-    /// Fully-diluted market cap `(FDMC) = price x max supply`. If max supply is null, `FDMC =
-    /// price x total supply`. If max supply and total supply are infinite or not available,
-    /// fully-diluted market cap shows `- -`.
-    pub fully_diluted_market_cap: Decimal,
-    pub last_updated: DateTime<Utc>,
+fn setup(listener: TcpListener, db_pool: PgPool) -> Result<Server, std::io::Error> {
+    let pool = Data::new(db_pool);
+    let server = HttpServer::new(move || {
+        App::new()
+            .route("/health_check", web::get().to(health_check::health_check))
+            .route("/crypto/map/", web::get().to(crypto_map::map))
+            .route("/crypto/listing/", web::get().to(crypto_listing::listing))
+            .app_data(pool.clone())
+    })
+    .listen(listener)?
+    .run();
+
+    Ok(server)
 }
