+++ crypto-forex/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cfx;
 pub mod config;
 pub mod database;
+pub mod model;
 pub mod route;
